requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
PyMuPDF
Pillow
pytesseract
python-multipart
Jinja2